LUCENE-3220: fix integer math division bug in gl2
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/flexscoring@1155438 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/test/org/apache/lucene/search/similarities/BasicModelG.java
@@ -26,7 +26,7 @@ import static org.apache.lucene.search.similarities.EasySimilarity.log2;
 public class BasicModelG extends BasicModel {
   @Override
   public final float score(EasyStats stats, float tfn) {
-    double lambda = stats.getTotalTermFreq() / stats.getNumberOfDocuments();
+    double lambda = stats.getTotalTermFreq() / (double) stats.getNumberOfDocuments();
     // -log(1 / (lambda + 1)) -> log(lambda + 1)
     return (float)(log2(lambda + 1) + tfn * log2((1 + lambda) / lambda));
   }
